+++ CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - New function `getMissingBackends` for `Logs`
 - New property `federation:backends` added to the array returned by `validateProcess`
+- New functions in `Connection` for paginating through lists:
+  - `paginateProcesses`
+  - `paginateCollections`
+  - `paginateJobs`
+  - `paginateFiles`
+  - `paginateUserProcesses`
+  - `paginateServices`
 
 ## [2.6.0] - 2024-07-11